+++ example/lib/main.dart
@@ -48,9 +48,12 @@ class _MyHomePageState extends State<MyHomePage> {
   @override
   Widget build(BuildContext context) {
     void _handleCheckBoxState({bool updateState = true}) {
-      var selectedItem = _popupBuilderKey.currentState?.popupGetSelectedItems ?? [];
-      var isAllSelected = _popupBuilderKey.currentState?.popupIsAllItemSelected ?? false;
-      _popupBuilderSelection = selectedItem.isEmpty ? false : (isAllSelected ? true : null);
+      var selectedItem =
+          _popupBuilderKey.currentState?.popupGetSelectedItems ?? [];
+      var isAllSelected =
+          _popupBuilderKey.currentState?.popupIsAllItemSelected ?? false;
+      _popupBuilderSelection =
+          selectedItem.isEmpty ? false : (isAllSelected ? true : null);
 
       if (updateState) setState(() {});
     }
@@ -74,7 +77,10 @@ class _MyHomePageState extends State<MyHomePage> {
                 children: [
                   Expanded(
                     child: DropdownSearch<int>(
-                      items: [1, 2, 3, 4, 5, 6, 7],
+                      dropdownDecoratorProps: DropDownDecoratorProps(
+                        showSuffixIcon: false,
+                      ),
+                      items: Iterable<int>.generate(100).toList(),
                     ),
                   ),
                   Padding(padding: EdgeInsets.all(4)),
@@ -113,7 +119,8 @@ class _MyHomePageState extends State<MyHomePage> {
                               child: MaterialButton(
                                 child: Text('OK'),
                                 onPressed: () {
-                                  _popupCustomValidationKey.currentState?.popupOnValidate();
+                                  _popupCustomValidationKey.currentState
+                                      ?.popupOnValidate();
                                 },
                               ),
                             ),
@@ -137,7 +144,9 @@ class _MyHomePageState extends State<MyHomePage> {
                         ),
                       ),
                       popupProps: PopupProps.bottomSheet(
-                          bottomSheetProps: BottomSheetProps(elevation: 16, backgroundColor: Color(0xFFAADCEE))),
+                          bottomSheetProps: BottomSheetProps(
+                              elevation: 16,
+                              backgroundColor: Color(0xFFAADCEE))),
                     ),
                   ),
                   Padding(padding: EdgeInsets.all(4)),
@@ -177,7 +186,9 @@ class _MyHomePageState extends State<MyHomePage> {
                         favoriteItemProps: FavoriteItemProps(
                           showFavoriteItems: true,
                           favoriteItems: (us) {
-                            return us.where((e) => e.name.contains("Mrs")).toList();
+                            return us
+                                .where((e) => e.name.contains("Mrs"))
+                                .toList();
                           },
                         ),
                       ),
@@ -194,11 +205,14 @@ class _MyHomePageState extends State<MyHomePage> {
                         favoriteItemProps: FavoriteItemProps(
                           showFavoriteItems: true,
                           favoriteItems: (us) {
-                            return us.where((e) => e.name.contains("Mrs")).toList();
+                            return us
+                                .where((e) => e.name.contains("Mrs"))
+                                .toList();
                           },
                           favoriteItemBuilder: (context, item, isSelected) {
                             return Container(
-                              padding: EdgeInsets.symmetric(horizontal: 8, vertical: 6),
+                              padding: EdgeInsets.symmetric(
+                                  horizontal: 8, vertical: 6),
                               decoration: BoxDecoration(
                                   border: Border.all(color: Colors.grey),
                                   borderRadius: BorderRadius.circular(10),
@@ -211,7 +225,9 @@ class _MyHomePageState extends State<MyHomePage> {
                                     style: TextStyle(color: Colors.indigo),
                                   ),
                                   Padding(padding: EdgeInsets.only(left: 8)),
-                                  isSelected ? Icon(Icons.check_box_outlined) : SizedBox.shrink(),
+                                  isSelected
+                                      ? Icon(Icons.check_box_outlined)
+                                      : SizedBox.shrink(),
                                 ],
                               ),
                             );
@@ -249,7 +265,8 @@ class _MyHomePageState extends State<MyHomePage> {
                       validator: (List<int>? items) {
                         if (items == null || items.isEmpty)
                           return 'required filed';
-                        else if (items.length > 3) return 'only 1 to 3 items are allowed';
+                        else if (items.length > 3)
+                          return 'only 1 to 3 items are allowed';
                         return null;
                       },
                     ),
@@ -310,8 +327,11 @@ class _MyHomePageState extends State<MyHomePage> {
                             isSelected: _popupBuilderSelection,
                             onChanged: (v) {
                               if (v == true)
-                                _popupBuilderKey.currentState!.popupSelectAllItems();
-                              else if (v == false) _popupBuilderKey.currentState!.popupDeselectAllItems();
+                                _popupBuilderKey.currentState!
+                                    .popupSelectAllItems();
+                              else if (v == false)
+                                _popupBuilderKey.currentState!
+                                    .popupDeselectAllItems();
                               _handleCheckBoxState();
                             },
                           );
@@ -340,7 +360,8 @@ class _MyHomePageState extends State<MyHomePage> {
                                     child: OutlinedButton(
                                       onPressed: () {
                                         // How should I unselect all items in the list?
-                                        _multiKey.currentState?.closeDropDownSearch();
+                                        _multiKey.currentState
+                                            ?.closeDropDownSearch();
                                       },
                                       child: const Text('Cancel'),
                                     ),
@@ -350,7 +371,8 @@ class _MyHomePageState extends State<MyHomePage> {
                                     child: OutlinedButton(
                                       onPressed: () {
                                         // How should I select all items in the list?
-                                        _multiKey.currentState?.popupSelectAllItems();
+                                        _multiKey.currentState
+                                            ?.popupSelectAllItems();
                                       },
                                       child: const Text('All'),
                                     ),
@@ -360,7 +382,8 @@ class _MyHomePageState extends State<MyHomePage> {
                                     child: OutlinedButton(
                                       onPressed: () {
                                         // How should I unselect all items in the list?
-                                        _multiKey.currentState?.popupDeselectAllItems();
+                                        _multiKey.currentState
+                                            ?.popupDeselectAllItems();
                                       },
                                       child: const Text('None'),
                                     ),
@@ -404,12 +427,14 @@ class _MyHomePageState extends State<MyHomePage> {
                           ),
                         ),
                       ),
-                      compareFn: (item, selectedItem) => item.id == selectedItem.id,
+                      compareFn: (item, selectedItem) =>
+                          item.id == selectedItem.id,
                       dropdownDecoratorProps: DropDownDecoratorProps(
                         dropdownSearchDecoration: InputDecoration(
                           labelText: 'Users *',
                           filled: true,
-                          fillColor: Theme.of(context).inputDecorationTheme.fillColor,
+                          fillColor:
+                              Theme.of(context).inputDecorationTheme.fillColor,
                         ),
                       ),
                       dropdownBuilder: _customDropDownExampleMultiSelection,
@@ -431,7 +456,8 @@ class _MyHomePageState extends State<MyHomePage> {
                         dropdownSearchDecoration: InputDecoration(
                           labelText: 'User *',
                           filled: true,
-                          fillColor: Theme.of(context).inputDecorationTheme.fillColor,
+                          fillColor:
+                              Theme.of(context).inputDecorationTheme.fillColor,
                         ),
                       ),
                     ),
@@ -563,10 +589,13 @@ class _MyHomePageState extends State<MyHomePage> {
                                 children: item.subLevel
                                     .map(
                                       (e) => ListTile(
-                                        selected: myKey.currentState?.getSelectedItem?.level1 == e.level1,
+                                        selected: myKey.currentState
+                                                ?.getSelectedItem?.level1 ==
+                                            e.level1,
                                         title: Text(e.level1),
                                         onTap: () {
-                                          myKey.currentState?.popupValidate([e]);
+                                          myKey.currentState
+                                              ?.popupValidate([e]);
                                         },
                                       ),
                                     )
@@ -586,7 +615,8 @@ class _MyHomePageState extends State<MyHomePage> {
     );
   }
 
-  Widget _customDropDownExampleMultiSelection(BuildContext context, List<UserModel> selectedItems) {
+  Widget _customDropDownExampleMultiSelection(
+      BuildContext context, List<UserModel> selectedItems) {
     if (selectedItems.isEmpty) {
       return ListTile(
         contentPadding: EdgeInsets.all(0),
@@ -616,7 +646,8 @@ class _MyHomePageState extends State<MyHomePage> {
     );
   }
 
-  Widget _customPopupItemBuilderExample2(BuildContext context, UserModel item, bool isSelected) {
+  Widget _customPopupItemBuilderExample2(
+      BuildContext context, UserModel item, bool isSelected) {
     return Container(
       margin: EdgeInsets.symmetric(horizontal: 8),
       decoration: !isSelected