+++ lib/src/properties/dropdown_button_props.dart
@@ -3,7 +3,8 @@ import 'package:flutter/material.dart';
 
 class DropdownButtonProps extends IconButtonProps {
   const DropdownButtonProps({
-    super.icon = const Icon(Icons.arrow_drop_down, size: 24),
+    super.icon = 
+     const Icon(Icons.arrow_drop_down, size: 24),
     super.isVisible = true,
     super.iconSize = 24.0,
     super.visualDensity,