+++ lib/src/properties/dropdown_decorator_props.dart
@@ -5,7 +5,7 @@ class DropDownDecoratorProps {
   final TextStyle? baseStyle;
   final TextAlign? textAlign;
   final TextAlignVertical? textAlignVertical;
-
+  final bool showSuffixIcon;
   /// Defaults to the value of [ThemeData.splashColor] or the nearest [Theme].
   final Color? searchSplashColor;
 
@@ -36,5 +36,6 @@ class DropDownDecoratorProps {
     this.autofocus = false,
     this.canRequestFocus = true,
     this.focusNode,
+    this.showSuffixIcon=true,
   });
 }